back/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==22.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
psutil==7.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.3
pandas==2.2.2
openpyxl==3.1.2
xlsxwriter==3.2.0
xlwt==1.3.0